refactor(mpp): extract frame codec into fork_portable submodule (R14)
R14 of the gap-closing plan. Make the "could-move-to-fork" boundary
explicit at the file level by splitting transport.rs's frame codec into
a new `mpp/fork_portable/frame.rs`. The new submodule has zero
`pgrx::*` or PG-specific references — promotion to a real fork PR
(paradedb/datafusion-distributed) would lift it as-is.

Scope (deliberately narrow):

- Move `MppFrameHeader`, `MppFrameKind`, `encode_frame_into`,
  `encode_eof_frame_into`, `decode_frame`, `MPP_FRAME_MAGIC`,
  `MPP_FRAME_HEADER_SIZE`, `FRAME_KIND_MASK`, and the size-assertion
  `const _: ()` into `fork_portable/frame.rs`.
- Move the 8 codec unit tests (`frame_round_trips_*`,
  `frame_rejects_*`, `frame_eof_with_payload_is_rejected`,
  `codec_round_trips_many_batch_sizes`) and the `sample_batch` helper
  into the same module's `#[cfg(test)] mod tests`.
- New `fork_portable/mod.rs` documents the boundary, lists what's
  NOT moved (cooperative-spin `MppSender` in `transport.rs` stays
  PG-bound), and explicitly flags the channel-buffer infrastructure
  (`DrainBuffer`, `DrainHandle`, etc.) as structurally portable but
  out-of-scope for this PR because the existing ~850-LOC test module
  in `transport.rs` reaches into private fields via `cfg(test)`
  inherent impls.

Why scope down: the test-coupling for `DrainBuffer`/`DrainHandle`
needs a parallel relocation (either move ~850 LOC of tests to
`fork_portable/channel.rs` or expose private fields via `cfg(test)`
helpers). Either path is invasive. R14 deliberately ships the
clearest-boundary piece first; a follow-up can move the channel infra
if the fork ever needs those primitives.

Visibility: items previously `pub(super)` in transport.rs become
`pub(in crate::postgres::customscan::mpp)` in `fork_portable/frame.rs`
to preserve the "visible across mpp tree" reach. The single `pub` type
(`MppFrameHeader`) stays `pub`.

Imports updated:
- `glue.rs`: now imports `MppFrameHeader` from
  `mpp::fork_portable::frame` instead of `mpp::transport`.
- `exec_worker.rs`: same.
- `runtime.rs`: dead intra-doc link updated to the new path.

No behavior change. All 4 MPP regress tests pass; the moved codec
tests (8 of them) run from their new home.
