+++ pg_search/src/postgres/customscan/mpp/mod.rs
@@ -27,6 +27,7 @@
 
 pub mod dsm;
 pub mod exec_worker;
+pub mod fork_portable;
 pub mod glue;
 pub mod host;
 pub mod mesh;
